+++ modules/alloy/src/test/ts/browser/behaviour/docking/DockingTest.ts
@@ -237,7 +237,7 @@ describe('browser.alloy.behaviour.docking.DockingTest', () => {
       assertInitialStructures('After Docking.refresh', { static: staticBox, absolute: absoluteBox });
     });
 
-    it('External Scroll event', () => {
+    it('External Scroll event', async () => {
       const store = hook.store();
       const component = hook.component();
 
@@ -264,10 +264,12 @@ describe('browser.alloy.behaviour.docking.DockingTest', () => {
       // Now, broadcast an external scroll event to the mothership, and check that it triggers
       // refresh and puts the box back where it should be.
       hook.gui().broadcastEvent(SystemEvents.externalElementScroll(), { } as any);
-      assertInitialStructures(
-        'After broadcasting external scroll event',
-        { static: staticBox, absolute: absoluteBox }
-      );
+      await Waiter.pTryUntil('Waited for initial structure', () => {
+        assertInitialStructures(
+          'After broadcasting external scroll event',
+          { static: staticBox, absolute: absoluteBox }
+        );
+      });
     });
   });
 });